requirements.txt
streamlit
face_recognition
opencv-python
numpy
pandas
Pillow 